+++ tests/test_modules.py
@@ -109,6 +109,7 @@ class TestModules(unittest.TestCase):
     )
     from .modules.module_test import (
         test_modules_test_check_inputs,
+        test_modules_test_no_modules_dir,
         test_modules_test_no_name_no_prompts,
     )
     from .modules.remove import (